Implement Circle, a circular data structure
Circle is a circular data structure.

It's implemented using a map where every value is a three-element tuple
containing the id of the previous item, the value of the current item
and then the id of the next item.

There are also meta-keys like size, current and counter that are kept up
to date automatically as long as the built in API functions are used.